l'interface se met à jour correctement (données factices)

+++ Contract/src/main/java/org/example/IFridgeModel.java
@@ -3,7 +3,7 @@ package org.example;
 import java.util.Observer;
 
 public interface IFridgeModel  {
-    public void setProps(int internTemp, int externTemp, float humidity);
+    public void setProps(int internTemp, int externTemp, double humidity);
     public void setParams(IFridgeParams params);
     public IFridgeParams getParams();
     public void addObserver(Observer args);
